Various Updates
Made it so users could not have the same username. This was needed because the application tracks what username submitted interest in an event. If multiple users shared a username, some users may not be able to submit interest.

Also cleaned up code.

Also cleaned up the design of the sign-up page.

Also added popup to show if an account was successfully created.